requirements.txt
backcall==0.1.0
boto3==1.9.39
botocore==1.12.39
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.0
decorator==4.3.0
dj-database-url==0.5.0
Django==2.1.3
django-cors-headers==2.4.0
django-dotenv==1.4.2
django-extensions==2.1.3
djangorestframework==3.9.0
docutils==0.14
gunicorn==19.9.0
idna==2.7
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
jmespath==0.9.3
parso==0.3.1
pickleshare==0.7.5
prompt-toolkit==2.0.7
psycopg2==2.7.5
Pygments==2.2.0
PyJWT==1.7.1
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.0
s3transfer==0.1.13
six==1.11.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
